refactor: 기존 예외 응답을 ExceptionResponse로 교체하여 공통 예외 응답을 통합 처리 [#back-59]
